+++ eras/conway/impl/src/Cardano/Ledger/Conway/Genesis.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE UndecidableInstances #-}
@@ -13,30 +14,66 @@ import Cardano.Ledger.Binary (
   DecCBOR (..),
   EncCBOR (..),
  )
+import Cardano.Ledger.Conway.Era (ConwayEra)
+import Cardano.Ledger.Conway.Governance
 import Cardano.Ledger.Conway.PParams (UpgradeConwayPParams)
 import Cardano.Ledger.Crypto (Crypto)
-import Data.Aeson (FromJSON (..), ToJSON)
-import Data.Aeson.Types (ToJSON (..))
+import Data.Aeson (
+  FromJSON (..),
+  KeyValue (..),
+  ToJSON (..),
+  object,
+  pairs,
+  withObject,
+  (.:),
+ )
+import Data.Default.Class (Default (def))
 import Data.Functor.Identity (Identity)
 import GHC.Generics (Generic)
 import NoThunks.Class (NoThunks)
 
-newtype ConwayGenesis c = ConwayGenesis
-  { cgUpgradePParams :: UpgradeConwayPParams Identity
+data ConwayGenesis c = ConwayGenesis
+  { cgUpgradePParams :: !(UpgradeConwayPParams Identity)
+  , cgConstitution :: !(Constitution (ConwayEra c))
+  , cgCommittee :: !(Committee (ConwayEra c))
   }
   deriving (Eq, Generic, Show)
 
-instance NoThunks (ConwayGenesis c)
+instance Crypto c => NoThunks (ConwayGenesis c)
 
 -- | Genesis are always encoded with the version of era they are defined in.
 instance Crypto c => DecCBOR (ConwayGenesis c) where
-  decCBOR = ConwayGenesis <$> decCBOR
+  decCBOR =
+    ConwayGenesis
+      <$> decCBOR
+      <*> decCBOR
+      <*> decCBOR
 
 instance Crypto c => EncCBOR (ConwayGenesis c) where
-  encCBOR (ConwayGenesis x) = encCBOR x
+  encCBOR (ConwayGenesis pparams constitution committee) =
+    encCBOR pparams
+      <> encCBOR constitution
+      <> encCBOR committee
 
 instance Crypto c => ToJSON (ConwayGenesis c) where
-  toJSON ConwayGenesis {..} = toJSON cgUpgradePParams
+  toJSON = object . toConwayGenesisPairs
+  toEncoding = pairs . mconcat . toConwayGenesisPairs
 
 instance Crypto c => FromJSON (ConwayGenesis c) where
-  parseJSON x = ConwayGenesis <$> parseJSON x
+  parseJSON =
+    withObject "ConwayGenesis" $ \obj ->
+      ConwayGenesis
+        <$> obj .: "upgradeProtocolParams"
+        <*> obj .: "constitution"
+        <*> obj .: "committee"
+
+toConwayGenesisPairs :: (Crypto c, KeyValue a) => ConwayGenesis c -> [a]
+toConwayGenesisPairs cg@(ConwayGenesis _ _ _) =
+  let ConwayGenesis {..} = cg
+   in [ "upgradeProtocolParams" .= cgUpgradePParams
+      , "constitution" .= cgConstitution
+      , "committee" .= cgCommittee
+      ]
+
+instance Crypto c => Default (ConwayGenesis c) where
+  def = ConwayGenesis def def def